refactor vclib.svn (prepare to add svn commandline client based driver)
  * move commonly used class and functions not using swig Python binding
    into svn_common.py

  * move commonly used functions using swig Python binding and API version
    check into common_swig_py.py

  * create common super class of LocalSubversionRepository and
    RemoteSubversionRepository (and may be CommandLineSubversionRepository)
    into common.py

  * move _getpath() instance method into normal function in common.py.
    (I think it can be provided by more upper module, though)

  * prepare to port to newer Python (keep Python 2.4 support)
    - use floor division operator in _datestr_to_date()
    - use isinstance() istead of extract type with type()
      in _getrev(), _rev2optrev(rev)
